feat: Change Posts.jsx to postTransformers.js
